Corrected minor bug seen on sonar

+++ src/main/java/es/upm/miw/forge/ecp2/Point.java
@@ -30,7 +30,7 @@ public class Point {
     }
     
     public double module() {
-        return Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z);
+        return Math.sqrt(this.x * this.x + this.y * this.y + (double)this.z * this.z);
     }
 
     public double phase() {
